Misc restructuring in async-opcua-codegen
This crate has become pretty messy. This is a first round of some
non-functional refactoring. Generally just moving functions and types
around a bit, moving some common stuff to utils, and splitting some
large files.
